Got CDB/PDB figured out
So this is an interesting run.

See: https://dba.stackexchange.com/questions/186915/creating-oracle-12c-r2-user-with-default-collation-binary-ci-fails

When creating the user, you must FIRST set the container to the PDB you
are creating the user in. If you do not, it will still think that the
max_string_size is STANDARD.

Upped the memory from 2048 to 4096
Increased the CPU count to 2

Pulled run of utlrp.sql from execution - was taking significant amount
of time to run.